Core/Scripts : Resolve type truncation and signed mismatch warnings
- Fixed float-to-bool truncations in MoveJump calls across 16 script files and resolved double-to-float precision issues in SmartAI and boss scripts.
- Standardized signed/unsigned comparisons across core systems (Player, Unit, ConditionMgr) and various Legion/Argus scripts.
- Refactored risky int8 and boolean comparisons in scenario scripts for better type safety.
- Math Cleanup: Applied static_casts and 'f' suffixes to M_PI calculations and PET_FOLLOW_ANGLE constants.

+++ src/server/game/BattlePay/BattlePayMgr.cpp
@@ -448,7 +448,7 @@ void BattlepayManager::SendProductList()
         bool hideProductPrice = false;
         if (pInfo.DisplayInfo.has_value() && pInfo.DisplayInfo->Flags.has_value())
             hideProductPrice = *pInfo.DisplayInfo->Flags & BattlepayDisplayInfoFlag::HidePrice;
-        bool hasEnoughTokens = tokenBalance >= product.CurrentPriceFixedPoint;
+        bool hasEnoughTokens = tokenBalance >= static_cast<int64>(product.CurrentPriceFixedPoint);
 
         response.ProductList.ProductInfo.emplace_back(pInfo);
